Move agenda-view component to page directory
