+++ libs/page/agenda-view/.babelrc
