+++ libs/page/agenda-view/eslint.config.js
@@ -1,5 +1,5 @@
 const nx = require('@nx/eslint-plugin');
-const baseConfig = require('../../eslint.config.js');
+const baseConfig = require('../../../eslint.config.js');
 
 module.exports = [
   ...baseConfig,
