+++ libs/page/agenda-view/jest.config.ts
@@ -1,10 +1,10 @@
 export default {
   displayName: 'agenda-view',
-  preset: '../../jest.preset.js',
+  preset: '../../../jest.preset.js',
   transform: {
     '^(?!.*\\.(js|jsx|ts|tsx|css|json)$)': '@nx/react/plugins/jest',
     '^.+\\.[tj]sx?$': ['babel-jest', { presets: ['@nx/react/babel'] }],
   },
   moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
-  coverageDirectory: '../../coverage/libs/agenda-view',
+  coverageDirectory: '../../../coverage/libs/page/agenda-view',
 };
