+++ libs/page/agenda-view/project.json
@@ -1,7 +1,7 @@
 {
   "name": "agenda-view",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/agenda-view/src",
+  "$schema": "../../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "libs/page/agenda-view/src",
   "projectType": "library",
   "tags": [],
   "// targets": "to see all targets run: nx show project agenda-view --web",
