+++ libs/page/agenda-view/src/lib/agenda-view.tsx
