+++ libs/page/agenda-view/tsconfig.lib.json
